+++ ios/ScreenShareAccPackKit/CHANGELOG.md
@@ -4,6 +4,19 @@ The changelog for `screensharing-annotation-acc-pack` iOS.
 
 --------------------------------------
 
+2.0.0
+-----
+
+### Enhancements
+- Remove shared instance so developers can control it.
+
+1.1.6
+-----
+
+### Enhancements
+
+- Introduce `isRemoteAudioAvailable` and `isRemoteVideoAvailable` to have more audio and video control.
+
 1.1.5
 -----